conversion error fixed

+++ src/Wordle.cpp
@@ -18,7 +18,7 @@ void Wordle::initHashmap() {
 }
 
 sf::Vector2i Wordle::findWord(const std::string& word) {
-    for (int i = 0; i < word.length(); i++){
+    for (int i = 0; i < (int)word.length(); i++){
         for (int j = word.length()-i; j >= 1; j--){
 
             std::string substring = word.substr(i, j);
